update: /common/handleDisconnect 조건 수정 (undefined일때 예기치않은 오류 발생)

+++ src/meeting/meeting.gateway.ts
@@ -43,9 +43,9 @@ export class MeetingGateway
   private connectedUsers: { [nickname: string]: Socket } = {} // nickname: socketId 형태로 변경
   private connectedSockets: { [socketId: string]: string } = {} // socketId: nickname 형태로 변경
   private cupidFlag: Map<string, boolean> = new Map()
-  private lastCupidFlag: Map<string,boolean> = new Map()
+  private lastCupidFlag: Map<string, boolean> = new Map()
 
-  private acceptanceStatus: Record<string, boolean> = {};
+  private acceptanceStatus: Record<string, boolean> = {}
   afterInit(server: Server) {
     this.meetingService.server = server
     console.log('WebSocket initialized')
@@ -56,7 +56,8 @@ export class MeetingGateway
   handleDisconnect(client: Socket) {
     const sessions = this.meetingService.getSessions()
     const participantName = this.connectedSockets[client.id]
-    if (!sessions.length) {
+    const user = client['user']
+    if (!sessions.length && user) {
       const gender = client['user'].gender
       this.queueService.removeParticipant(participantName, gender)
     }
@@ -74,9 +75,9 @@ export class MeetingGateway
     delete this.connectedUsers[participantName]
     this.roomid.delete(participantName)
 
-    const nickname = this.connectedSockets[client.id];
+    const nickname = this.connectedSockets[client.id]
     if (nickname in this.acceptanceStatus) {
-      delete this.acceptanceStatus[nickname];
+      delete this.acceptanceStatus[nickname]
     }
   }
 
@@ -107,7 +108,6 @@ export class MeetingGateway
         )
         this.roomid.delete(participantName)
       }
-
 
       const { sessionName, readyMales, readyFemales } =
         await this.queueService.handleJoinQueue(participantName, client, gender)
@@ -206,7 +206,7 @@ export class MeetingGateway
           })
           this.cupidFlag.set(sessionName, true)
         }
-        this.meetingService.removeChooseData(sessionName);
+        this.meetingService.removeChooseData(sessionName)
       }
     } else {
       console.error('세션에러입니다')
@@ -273,43 +273,40 @@ export class MeetingGateway
     })
   }
 
-
   // 1. 10초 이내에 '1대1화상채팅하기' 버튼을 누르지 않으면 비활성
   // 2. 성공적으로 '1대1화상채팅하기' 버튼을 눌렀을 경우 클라이언트 -> 서버(Event : chooseCam)
   @SubscribeMessage('lastChoose')
-  handleChooseCam(client: Socket, payload: { sender : string; receiver: string })
-  { // 서버 입장에서 소켓이 존재하는 방을 찾기 위함
-    const { sender, receiver }  = payload
+  handleChooseCam(
+    client: Socket,
+    payload: { sender: string; receiver: string },
+  ) {
+    // 서버 입장에서 소켓이 존재하는 방을 찾기 위함
+    const { sender, receiver } = payload
     const sessionName = this.roomid.get(sender)
     // 기존 정보가 있다면 새롭게 변형해서 저장할 수 있음
     if (sessionName) {
-      this.meetingService.storeChoose(
-        sessionName,
-        sender,
-        receiver,
-      )
+      this.meetingService.storeChoose(sessionName, sender, receiver)
       // 방과 일치하는 매칭결과 정보 가져오기
-      const chooseData = this.meetingService.getChooseData(sessionName);
-      if ( chooseData.length === 6) {
+      const chooseData = this.meetingService.getChooseData(sessionName)
+      if (chooseData.length === 6) {
         // 방과 일치하는 참여자 정보 가져오기
-        const participant = this.meetingService.getParticipants(sessionName);
+        const participant = this.meetingService.getParticipants(sessionName)
         // 매칭된 쌍의 정보를 가지고 있음
         // [
         // { pair: [ 'Alice', 'Bob' ] },
         // { pair: [ 'Charlie', 'David' ] },
         // { pair: [ 'Eve', 'Frank' ] }
         // ]
-        const matches = this.meetingService.findMatchingPairs(sessionName);
+        const matches = this.meetingService.findMatchingPairs(sessionName)
 
         if (this.lastCupidFlag.get(sessionName) == undefined) {
-          participant.forEach( ({socket, name }) => {
-
+          participant.forEach(({ socket, name }) => {
             const matchedPair = matches.find(elem => elem.pair.includes(name))
             if (matchedPair) {
-              const partner = matchedPair.pair.find( elem => elem !== name)
-              this.server.to(socket.id).emit('matching', { lover : partner })
+              const partner = matchedPair.pair.find(elem => elem !== name)
+              this.server.to(socket.id).emit('matching', { lover: partner })
             } else {
-              this.server.to(socket.id).emit('matching', { lover : '0'})
+              this.server.to(socket.id).emit('matching', { lover: '0' })
             }
 
             this.server.to(socket.id).emit('lastChooseResult', chooseData)
@@ -323,32 +320,46 @@ export class MeetingGateway
   }
 
   @SubscribeMessage('moveToPrivateRoom')
-  async handleMoveToPrivateRoom(client: Socket, payload : {sessionName:string; myName:string; partnerName:string})
-  { 
-    
-    const { sessionName, myName, partnerName } = payload;
-    const participant = this.meetingService.getParticipants(sessionName);
+  async handleMoveToPrivateRoom(
+    client: Socket,
+    payload: { sessionName: string; myName: string; partnerName: string },
+  ) {
+    const { sessionName, myName, partnerName } = payload
+    const participant = this.meetingService.getParticipants(sessionName)
     if (this.acceptanceStatus[partnerName] === true) {
-      const newSessionName = `${myName}-${partnerName}`;
-      const newSession = await this.meetingService.createSession(newSessionName);
+      const newSessionName = `${myName}-${partnerName}`
+      const newSession = await this.meetingService.createSession(newSessionName)
 
-      const partner = await participant.find(participant => participant.name === partnerName )
-      this.meetingService.addParticipant(newSessionName, myName, client);
-      this.meetingService.addParticipant(newSessionName, partnerName, partner.socket);
+      const partner = await participant.find(
+        participant => participant.name === partnerName,
+      )
+      this.meetingService.addParticipant(newSessionName, myName, client)
+      this.meetingService.addParticipant(
+        newSessionName,
+        partnerName,
+        partner.socket,
+      )
 
-      const enterToken = await this.meetingService.generateTokens(newSessionName);
+      const enterToken =
+        await this.meetingService.generateTokens(newSessionName)
 
-      const myToken = enterToken.find(elem => elem.participant === myName).token;
-      const partnerToken = enterToken.find(elem => elem.participant === partnerName).token;
+      const myToken = enterToken.find(elem => elem.participant === myName).token
+      const partnerToken = enterToken.find(
+        elem => elem.participant === partnerName,
+      ).token
 
       if (myToken && partnerToken) {
-        this.server.to(client.id).emit('choice', { sessionId: newSessionName, token: myToken });
-        this.server.to(partner.socket.id).emit('choice', { sessionId: newSessionName, token: partnerToken });
+        this.server
+          .to(client.id)
+          .emit('choice', { sessionId: newSessionName, token: myToken })
+        this.server
+          .to(partner.socket.id)
+          .emit('choice', { sessionId: newSessionName, token: partnerToken })
       } else {
-        console.error("방 생성 실패!");
+        console.error('방 생성 실패!')
       }
     } else {
-      this.acceptanceStatus[myName] = true;
+      this.acceptanceStatus[myName] = true
     }
   }
